+++ backend/pium/src/main/java/com/official/pium/service/NotificationService.java
@@ -1,11 +1,11 @@
-package com.official.pium.fcm;
+package com.official.pium.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class NotificationService {
