--- backend/pium/src/main/java/com/official/pium/service/dto/LoginRequest.java
@@ -1,20 +1,0 @@
-package com.official.pium.service.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class LoginRequest {
-
-    @NotNull
-    private String code;
-
-    private String deviceToken;
-}
