+++ backend/pium/src/main/resources/templates/admin/dict/create.html
@@ -100,8 +100,6 @@
             })
                 // .then(response => response.json())
                 .then(response => {
-                    console.log(response)
-                    console.log(response?.headers?.get("location"))
                     if(response?.headers?.get("location")) {
                         let location = response.headers.get("location");
                         window.location.replace(`/admin/dict/${location.split('/')[2]}`)
